Reorganize some sources
