+++ b9/CMakeLists.txt
@@ -1,12 +1,12 @@
 add_library(b9 SHARED
+	src/assemble.cpp
+	src/Compiler.cpp
+	src/deserialize.cpp
 	src/ExecutionContext.cpp
 	src/MethodBuilder.cpp
-	src/core.cpp
-	src/Compiler.cpp
 	src/primitives.cpp
 	src/serialize.cpp
-	src/deserialize.cpp
-	src/assemble.cpp
+	src/VirtualMachine.cpp
 )
 
 target_include_directories(b9
