+++ b9/src/serialize.cpp
@@ -5,7 +5,7 @@
 #include <vector>
 
 #include <b9/instructions.hpp>
-#include <b9/module.hpp>
+#include <b9/Module.hpp>
 #include <b9/serialize.hpp>
 
 namespace b9 {
