+++ b9disasm/b9disasm.cpp
@@ -2,7 +2,7 @@
 #include <fstream>
 
 #include <b9/deserialize.hpp>
-#include <b9/module.hpp>
+#include <b9/Module.hpp>
 
 using namespace b9;
